fix(vanilla): Cancel promise on unsubscribe
I'm seeing Errors like this in my React-App:

Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.

These occur both with Firefox and Chromium.

I think they are caused because a component is unmounted in the tick right between promise creation and resolution, so deleting the listener is not enough tho prevent the callback from being invoked.

The proposed change removes the errors.