fix: subscribers should not be called after unsubscribe
There is an edge case in the subscribe notification logic which can
cause a subscriber to be called after it has been unsubscribed.

To reproduce the issue:
  1. create a subscription "A"
  2. create a subscription "B"
  3. in the listener for "A", unsubscribe "B"

Expected result:
"B" is never called (assuming listeners are processed in the order they
were added)

Actual result:
"B" is called once, and then never again

The fix is to check whether the subscription has been unsubscribed
before firing the subscription's notification callback.