Allow passthrough of Redux devtools options (#713)
* Allow passthrough of Redux devtools options

The current 'devtools()' util in Valtio restricts the Redux-devtools option to the 'name' field, and a custom 'enabled' field. But it might be useful to pass other devtools options too to Redux-devtools. My current case, which motivated this, is that I'm having a large data set in the proxy store (as a ref() of course), and any change to the proxy causes a significant lag in redux-devtools. This has been performance measured in Chrome. The solution is to customize redux-devtools' serializer to skip the large data part, as I don't need to see that part. This way I can further debug other perf-related bottlenecks in my heavy app. During this effort, it seemed like Valtio unnecessarily restricted the redux-devtools options object, so this change only lifts this restriction.

* Fix lint error

* fix typing and use rest

---------

Co-authored-by: daishi <daishi@axlight.com>